+++ src/goray/montecarlo.go
@@ -64,7 +64,7 @@ func (h *Halton) Float() float { return float(h.Float64()) }
 // This function can also be used to generate Sobol sequences and Larcher &
 // Pillichshammer sequences.  The particular algorithm used is described in
 // "Efficient Multidimensional Sampling" by Alexander Keller.
-func VanDerCorput(bits, r uint32) float {
+func VanDerCorput(bits, r uint32) float64 {
 	const multRatio = 0.00000000023283064365386962890625
 
 	bits = bits<<16 | bits>>16
@@ -72,5 +72,5 @@ func VanDerCorput(bits, r uint32) float {
 	bits = bits&0x0f0f0f0f<<4 | bits&0xf0f0f0f0>>4
 	bits = bits&0x33333333<<2 | bits&0xcccccccc>>2
 	bits = bits&0x55555555<<1 | bits&0xaaaaaaaa>>1
-	return float(float64(bits^r) * multRatio)
+	return float64(bits^r) * multRatio
 }
